+++ api/handlers/config.go
@@ -50,6 +50,7 @@ type RESTHandler struct {
 	Description  string
 	GitTag       string
 	GitCommit    string
+	RepoURL      string
 	ConformsTo   []string
 	T            Template
 	StorageSvc   *s3.S3
@@ -77,6 +78,11 @@ func NewRESTHander(gitTag, gitCommit string) *RESTHandler {
 		log.Warn("env variable API_NAME not set")
 	}
 
+	repoURL, exist := os.LookupEnv("REPO_URL")
+	if !exist {
+		log.Warn("env variable REPO_URL not set")
+	}
+
 	// working with pointers here so as not to copy large templates, yamls, and ActiveJobs
 	config := RESTHandler{
 		Name:        apiName,
@@ -84,6 +90,7 @@ func NewRESTHander(gitTag, gitCommit string) *RESTHandler {
 		Description: "SEPEX - Service for Encapsulated Processes Execution. An OGC API - Processes compliant server for executing processes locally or on cloud at scale.",
 		GitTag:      gitTag,
 		GitCommit:   gitCommit,
+		RepoURL:     repoURL,
 		ConformsTo: []string{
 			"http://schemas.opengis.net/ogcapi/processes/part1/1.0/openapi/schemas/",
 			"http://www.opengis.net/spec/ogcapi-processes-1/1.0/conf/ogc-process-description",
@@ -112,9 +119,16 @@ func NewRESTHander(gitTag, gitCommit string) *RESTHandler {
 
 	// Read all the html templates
 	funcMap := template.FuncMap{
-		"prettyPrint": prettyPrint, // to pretty print JSONs for results and metadata
-		"lower":       strings.ToLower,
-		"upper":       strings.ToUpper,
+		"prettyPrint":   prettyPrint, // to pretty print JSONs for results and metadata
+		"lower":         strings.ToLower,
+		"upper":         strings.ToUpper,
+		"lastSegment": func(s string) string {
+			parts := strings.Split(strings.TrimSuffix(s, "/"), "/")
+			if len(parts) > 0 {
+				return parts[len(parts)-1]
+			}
+			return s
+		},
 	}
 
 	config.T = Template{