+++ api/jobs/docker_jobs.go
@@ -363,8 +363,10 @@ func (j *DockerJob) WriteMetaData() {
 		return
 	}
 
+	repoURL := os.Getenv("REPO_URL")
+
 	md := metaData{
-		Context:         "https://github.com/Dewberry/sepex/blob/main/context.jsonld",
+		Context:         fmt.Sprintf("%s/blob/main/context.jsonld", repoURL),
 		JobID:           j.UUID,
 		Process:         p,
 		Image:           i,